remove legacy routes
